third_party/WebKit/LayoutTests/fast/css/getComputedStyle/disallowed-url-serialization.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target">
<script>
var testCases = [
  ['-webkit-box-reflect', 'below 0px url(\"chrome:\") 0 fill / auto / 0px stretch'],
  ['-webkit-mask-box-image-source', 'url("chrome:")'],
  ['-webkit-mask-image', 'url("chrome:")'],
  ['background-image', 'url("chrome:")'],
  ['border-image-source', 'url("chrome:")'],
  ['content', 'url("chrome:")'],
  ['list-style-image', 'url("chrome:")'],
  ['shape-outside', 'url("chrome:")'],
];

test(() => {
  for ([property, value] of testCases) {
    target.style[property] = value;
  }
  var computedStyle = getComputedStyle(target);
  for ([property, value] of testCases) {
    assert_equals(computedStyle[property], value, property);
  }
}, 'Disallowed URL protocols should serialise as their original URL strings');
</script>
